Cleaned up/refactored. Took games out of team_test setup, instead inputting instance stat values. Added test methods for preseason win percent/bust
